detectorplagio/requirements.txt
Django==3.1.3
django-heroku==0.3.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
django-cors-headers==3.5.0
djangorestframework==3.12.2
gunicorn==20.0.4
whitenoise==5.2.0
wn==0.9.1
spacy==3.2.4
spacy-legacy==3.0.9
spacy-loggers==1.0.2
pt-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.2.0/pt_core_news_sm-3.2.0-py3-none-any.whl
reportlab==3.6.9
docx2txt~=0.8
pdfminer.six==20220524